[Finishes #167331853]Fix token decoding

+++ server/usingDB/controllers/Property.js
@@ -116,13 +116,13 @@ const Property = {
 
 
     async getType(req, res) {
-        const text = 'SELECT * FROM properties ';
+        const text = 'SELECT * FROM properties WHERE type = $1 ';
         try {
-            const { rows } = await db.query(text);
+            const { rows } = await db.query(text, [req.query.type]);
             if (!rows) {
                 return res.status(404).json({ 
                     status: 404,
-                    error : 'reflection not found' });
+                    error : 'property not found' });
             }
             return res.status(200).json({
                 status: 200,
@@ -147,22 +147,15 @@ const Property = {
 
 
     async update(req, res) {
-        const findOneQuery = 'SELECT * FROM properties WHERE id=$1 AND owner = $2';
         const updateOneQuery = `UPDATE properties
       SET price=$1 WHERE id=$2 AND owner = $3`;
         try {
-            const { rows } = await db.query(findOneQuery, [req.params.id, req.owner]);
-            if (!rows[0]) {
-                return res.status(404).json({ 
-                    status: 404,
-                    error: 'property not found' });
-            }
             const values = [
                 req.body.price || rows[0].price,
                 req.params.id,
                 res.locals.user.email
             ];
-            const response = await db.query(updateOneQuery, values);
+            const { rows } = await db.query(updateOneQuery, values);
             return res.status(200).json({
                 status: 200,
                 message: "successfully updated",
@@ -186,18 +179,16 @@ const Property = {
 
 
     async delete(req, res) {
-        const deleteQuery = 'DELETE FROM properties WHERE id=$1';
+        const deleteQuery = 'DELETE FROM properties WHERE id=$1 AND owner =$2';
         try {
-            const { rows } = await db.query(deleteQuery, [req.params.id]);
-            if (!rows[0]) {
-                return res.status(404).json({ 
-                    status: 404,
-                    error : 'property not found' });
-            }
-            return res.status(204).json({ 
-                status: 204,
-                message: 'property is deleted' 
+            const { rows }  = await db.query(deleteQuery, [req.params.id, req.owner]);
+            console.log(rows)
+            if(rows.length === 0){
+            return res.status(200).json({ 
+                status: 200,
+                message: 'property successfully deleted' 
             });
+        }
         } catch (error) {
             return res.status(400).json({
                 status: 400,
